+++ src/controllers/ShippingCategoriesController.php
@@ -129,17 +129,36 @@ class ShippingCategoriesController extends BaseShippingSettingsController
     /**
      * @throws HttpException
      */
-    public function actionDelete(): Response
+    public function actionDelete(): ?Response
     {
         $this->requirePostRequest();
-        $this->requireAcceptsJson();
 
-        $id = $this->request->getRequiredBodyParam('id');
+        $id = $this->request->getBodyParam('id');
+        $ids = $this->request->getBodyParam('ids');
 
-        if (!Plugin::getInstance()->getShippingCategories()->deleteShippingCategoryById($id)) {
+        if ((!$id && empty($ids)) || ($id && !empty($ids))) {
+            throw new BadRequestHttpException('id or ids must be specified.');
+        }
+
+        if ($id) {
+            // If it is just the one id we know it has come from an ajax request on the table
+            $this->requireAcceptsJson();
+            $ids = [$id];
+        }
+
+        $failedIds = [];
+        foreach ($ids as $id) {
+            if (!Plugin::getInstance()->getShippingCategories()->deleteShippingCategoryById($id)) {
+                $failedIds[] = $id;
+            }
+        }
+
+        if (!empty($failedIds)) {
+            // @TODO: re-word this message at next translations update
             return $this->asFailure(Craft::t('commerce', 'Could not delete shipping category'));
         }
 
+        // @TODO: re-word add better message in next translation update
         return $this->asSuccess();
     }